requirements.txt
requests
python-dotenv
plotly
kaleido
pandas
python-telegram-bot
schedule
pypresence # for discord status
httpx==0.24.1